requirements.txt
# === Core ===
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0

# === Google AI (Vertex / Gemini) ===
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.4.0

# === Database ===
supabase>=2.3.0
psycopg2-binary>=2.9.9

# === Bulk CSV processing ===
polars>=0.20.0

# === HTTP (bulk ZIP downloads) ===
httpx>=0.27.0

# === PDF ===
PyMuPDF>=1.23.0
